// api/generated/models/createTalentoPerfilDtoArea.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Unibac API
 * Autenticación: POST /auth/login fija la cookie HttpOnly access_token. En Swagger UI, ejecuta login en este mismo origen y luego las rutas protegidas enviarán la cookie automáticamente. Desde el frontend usa fetch/axios con credentials: include. Registro público: POST /auth/register-public solo si PUBLIC_REGISTRATION_ENABLED=true.
 * OpenAPI spec version: 1.0
 */

export type CreateTalentoPerfilDtoArea =
  (typeof CreateTalentoPerfilDtoArea)[keyof typeof CreateTalentoPerfilDtoArea];

export const CreateTalentoPerfilDtoArea = {
  MUSICA: "MUSICA",
  ARTES_PLASTICAS: "ARTES_PLASTICAS",
  DISENO: "DISENO",
  AUDIOVISUAL: "AUDIOVISUAL",
  ARTES_ESCENICAS: "ARTES_ESCENICAS",
} as const;